# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zbarber API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/sync:
    post:
      operationId: syncAuth
      tags: [auth]
      summary: Sync Clerk identity with local profile (JIT provisioning)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncAuthInput"
      responses:
        "200":
          description: Synced user and optional barber profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSyncResult"

  # ── USER SELF-SERVICE ────────────────────────────────────────────────
...
  /financing-requests:
    get:
      operationId: listFinancingRequests
      tags: [financing]
      summary: List financing requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, reviewing, approved, rejected]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Financing requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingListResponse"
    post:
      operationId: createFinancingRequest
      tags: [financing]
      summary: Submit financing request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancingInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingRequest"
  /financing-requests/{id}:
    get:
      operationId: getFinancingRequest
      tags: [financing]
      summary: Get financing request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financing request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingRequest"
    patch:
      operationId: updateFinancingStatus
      tags: [financing]
      summary: Update financing request status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancingStatusUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingRequest"

  # ── CONFERENCES ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [client, barber, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        locationUpdatedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Barber:
      type: object
      required: [id, userId, salonName, city, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        salonName:
          type: string
        bio:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        city:
          type: string
        neighborhood:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, awaiting_document, approved, rejected, suspended]
        profileViews:
          type: integer
        subscriptionPlanId:
          type: ["integer", "null"]
        rejectionReason:
          type: ["string", "null"]
        suspensionReason:
          type: ["string", "null"]
        firstValidatedAt:
          type: ["string", "null"]
          format: date-time
        documentUrl:
          type: ["string", "null"]
        documentSubmittedAt:
          type: ["string", "null"]
          format: date-time
        documentDeadline:
          type: ["string", "null"]
          format: date-time
        documentReviewNote:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    FinancingRequest:
      type: object
      required: [id, barberId, amount, purpose, status, createdAt, monthlyRevenue, yearsActive, repaymentMonths, documents]
      properties:
        id:
          type: integer
        barberId:
          type: integer
        barberName:
          type: ["string", "null"]
        amount:
          type: number
        purpose:
          type: string
          enum: [renovation, tools, products, other]
        description:
          type: string
        monthlyRevenue:
          type: number
        yearsActive:
          type: integer
        repaymentMonths:
          type: integer
        documents:
          type: array
          items:
            type: string
        idDocument:
          type: ["string", "null"]
        guarantorIdDocument:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        adminNote:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    FinancingInput:
      type: object
      required: [amount, purpose, description, monthlyRevenue, yearsActive, repaymentMonths, documents, idDocument, guarantorIdDocument]
      properties:
        amount:
          type: number
          minimum: 50000
          maximum: 5000000
        purpose:
          type: string
          enum: [renovation, tools, products, other]
        description:
          type: string
          minLength: 30
        monthlyRevenue:
          type: number
          minimum: 0
        yearsActive:
          type: integer
          minimum: 0
        repaymentMonths:
          type: integer
          minimum: 3
          maximum: 24
        documents:
          type: array
          minItems: 1
          items:
            type: string
        idDocument:
          type: string
        guarantorIdDocument:
          type: string
    FinancingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        adminNote:
          type: ["string", "null"]
    FinancingListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FinancingRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── CONFERENCE ────────────────────────────────────────────────────
...
    SyncAuthInput:
      type: object
      properties:
        role:
          type: string
          enum: [client, barber]
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        country:
          type: string
        avatarUrl:
          type: string
...
    AuthSyncResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        barber:
          allOf:
            - $ref: "#/components/schemas/Barber"
          nullable: true

    # ── BARBER SELF-SERVICE ────────────────────────────────────────────
...