# lib/api-zod/src/generated/types/financingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zbarber API
 * OpenAPI spec version: 0.1.0
 */
import type { FinancingInputPurpose } from './financingInputPurpose';

export interface FinancingInput {
  /**
     * @minimum 50000
     * @maximum 5000000
     */
  amount: number;
  purpose: FinancingInputPurpose;
  /** @minLength 30 */
  description: string;
  /** @minimum 0 */
  monthlyRevenue: number;
  /** @minimum 0 */
  yearsActive: number;
  /**
     * @minimum 3
     * @maximum 24
     */
  repaymentMonths: number;
  /** @minItems 1 */
  documents: string[];
  idDocument: string;
  guarantorIdDocument: string;
}