# lib/api-zod/src/generated/types/financingRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zbarber API
 * OpenAPI spec version: 0.1.0
 */
import type { FinancingRequestPurpose } from './financingRequestPurpose';
import type { FinancingRequestStatus } from './financingRequestStatus';

export interface FinancingRequest {
  id: number;
  barberId: number;
  /** @nullable */
  barberName?: string | null;
  amount: number;
  purpose: FinancingRequestPurpose;
  description?: string;
  monthlyRevenue: number;
  yearsActive: number;
  repaymentMonths: number;
  documents: string[];
  /** @nullable */
  idDocument?: string | null;
  /** @nullable */
  guarantorIdDocument?: string | null;
  status: FinancingRequestStatus;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  reviewedAt?: Date | null;
  createdAt: Date;
}